requirements.txt
numpy~=2.2.6
pandas~=2.3.3
opencv-python~=4.12.0.88
scikit-learn~=1.7.2
matplotlib
pyyaml
jupyter
scikit-image~=0.25.2
pillow~=12.0.0
joblib~=1.5.2
torch
torchvision
tensorflow